fix: Allow re-association after connection loss with eNB.
If no candidate eNB has more Rssi than rssi_min the UE is removed from the
eNB and the masterId is set to 0 (i.e. not connected). The candidateMasterRssi_
must not be set to 0 in this case as this implies a really good connection
RSSI to the cell with the masterId 0, which does not exist and is a placeholder
for 'not connected'. Setting the candidateMassterRssi_ to 0 would not
allow any re-assocation to any other eNB as no eNB will have  RSSSI > 0 (db/dbm).
Thus setting the candidateMasterRssi_ to a small value.